ref: refactored chat page
