--- src/components/chat.css
@@ -1,43 +1,0 @@
-.chat-main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  padding: 20px;
-  font-family: sans-serif;
-  border: 2px solid green;
-  height: 100vh;
-}
-
-.chat-main h2 {
-  margin-bottom: 20px;
-  font-size: 24px;
-}
-
-ul {
-  list-style: none;
-  padding: 0;
-  margin-bottom: 20px;
-}
-
-li {
-  margin-bottom: 10px;
-}
-
-li button {
-  padding: 10px 15px;
-  background-color: #f0f0f0;
-  border: 1px solid #ccc;
-  cursor: pointer;
-  font-size: 16px;
-  border-radius: 6px;
-  transition: background-color 0.2s ease;
-}
-
-li button:hover {
-  background-color: #e0e0e0;
-}
-
-li button:active {
-  border: 2px solid green;
-}
