+++ src/pages/chatpage.css
@@ -1,11 +1,52 @@
 .chat-page {
-    display: flex;
-    border: 2px solid blue;
-    background-color: rgb(33, 29, 219);
+  display: flex;
 }
 
 .chat-area {
-    width: 100%;
-    height: 100vh;
-    border: 3px solid yellow;
+  width: 100%;
+  height: 100vh;
+  /* border: 3px solid yellow; */
+}
+
+.chat-main {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  font-family: sans-serif;
+  /* border: 2px solid green; */
+  height: 100vh;
+}
+
+.chat-main h2 {
+  margin-bottom: 20px;
+  font-size: 24px;
+}
+
+ul {
+  list-style: none;
+  padding: 0;
+  margin-bottom: 20px;
+}
+
+li {
+  margin-bottom: 10px;
+}
+
+li button {
+  padding: 10px 15px;
+  background-color: #f0f0f0;
+  border: 1px solid #ccc;
+  cursor: pointer;
+  font-size: 16px;
+  border-radius: 6px;
+  transition: background-color 0.2s ease;
+}
+
+li button:hover {
+  background-color: #e0e0e0;
+}
+
+li button:active {
+  border: 2px solid green;
 }
